requirements.txt
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
eval_type_backport==0.2.0
Flask==3.0.3
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3